+++ SistemaFinancieroVentura/ejb/src/main/java/org/ventura/sistemafinanciero/service/AbstractService.java
@@ -1,5 +1,6 @@
 package org.ventura.sistemafinanciero.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.ventura.sistemafinanciero.exception.NonexistentEntityException;
@@ -9,15 +10,15 @@ public interface AbstractService<T> {
 
 	public void create(T t) throws PreexistingEntityException;
 	
-	public void update(int id, T t) throws NonexistentEntityException, PreexistingEntityException;
+	public void update(BigInteger id, T t) throws NonexistentEntityException, PreexistingEntityException;
 
-	public void delete(int id) throws NonexistentEntityException;
+	public void delete(BigInteger id) throws NonexistentEntityException;
 
-	public T findById(int id);
+	public T findById(BigInteger id);
 	
 	public List<T> findAll();
 
-	public List<T> findRange(int[] range);
+	public List<T> findRange(BigInteger[] range);
 
 	public int count();
