+++ SistemaFinancieroVentura/ejb/src/main/java/org/ventura/sistemafinanciero/service/TrabajadorService.java
@@ -1,5 +1,7 @@
 package org.ventura.sistemafinanciero.service;
 
+import java.math.BigInteger;
+
 import javax.ejb.Remote;
 
 import org.ventura.sistemafinanciero.entity.Agencia;
@@ -10,10 +12,10 @@ import org.ventura.sistemafinanciero.exception.NonexistentEntityException;
 @Remote
 public interface TrabajadorService extends AbstractService<Trabajador>{
 
-	public Trabajador findByUsuario(int idusuario);
+	public Trabajador findByUsuario(BigInteger idusuario);
 	
-	public Caja findByTrabajador(int idTrabajador) throws NonexistentEntityException;
+	public Caja findByTrabajador(BigInteger idTrabajador) throws NonexistentEntityException;
 	
-	public Agencia getAgencia(int idTrabajador) throws NonexistentEntityException;
+	public Agencia getAgencia(BigInteger idTrabajador) throws NonexistentEntityException;
 
 }
